backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
readme = ""
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.121.1",
    "mux-python>=5.1.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
